# src/OpenMetrc.Scraper/OpenApi/metrc-swagger-v2.yaml
swagger: 2.0
info:
  title: METRC API
  version: v2
paths:
...
  /patient-checkins/v2/{id}:
    delete:
      tags:
      - PatientCheckIns
      summary: ''
      description: '<b>Permissions Required</b>: ManagePatientsCheckIns'
      externalDocs:
        x-doc-key: delete_patient-checkins_v2_{id}
      operationId: DeletePatientCheckInById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to archive patient check-in.
        format: ''
      responses:
        200:
          description: Ok
  /patient-checkins/v2/:
    get:
      tags:
      - PatientCheckIns
      summary: ''
      description: '<b>Permissions Required</b>: ManagePatientsCheckIns'
      externalDocs:
        x-doc-key: get_patient-checkins_v2_
      operationId: GetPatientCheckIns
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of patient check-ins.
        format: ''
      - type: string
        name: checkinDateStart
        in: query
        description: The last modified start timestamp
        format: date-time
      - type: string
        name: checkinDateEnd
        in: query
        description: The last modified end timestamp
        format: date-time
      responses:
        200:
          description: Ok
    post:
      tags:
      - PatientCheckIns
      summary: ''
      description: '<b>Permissions Required</b>: ManagePatientsCheckIns'
      externalDocs:
        x-doc-key: post_patient-checkins_v2_
      operationId: PostPatientCheckIns
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record the list of patient check-ins.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - PatientCheckIns
      summary: ''
      description: '<b>Permissions Required</b>: ManagePatientsCheckIns'
      externalDocs:
        x-doc-key: put_patient-checkins_v2_
      operationId: PutPatientCheckIns
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of patient check-ins.
        format: ''
      responses:
        200:
          description: Ok
...
  /patients/v2/{id}:
    delete:
      tags:
      - Patients
      summary: ''
      description: '<b>Permissions Required</b>: Manage Patients'
      externalDocs:
        x-doc-key: delete_patients_v2_{id}
      operationId: DeletePatientById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to delete the patient.
        format: ''
      responses:
        200:
          description: Ok
    get:
      tags:
      - Patients
      summary: ''
      description: '<b>Permissions Required</b>: Manage Patients'
      externalDocs:
        x-doc-key: get_patients_v2_{id}
      operationId: GetPatientById
      parameters:
      - type: integer
        name: id
        in: path
        required: true
        description: Route parameter id
        format: int64
      - type: string
        name: licenseNumber
        in: query
        description: If specified, the Patient will be validated against the specified License Number. If not specified, the Patient will be validated against all of the User's current Facilities. Please note that if the Patient is not valid for the specified License Number, a 401 Unauthorized status will be returned.
        format: ''
      responses:
        200:
          description: Ok
  /patients/v2/active:
    get:
      tags:
      - Patients
      summary: ''
      description: '<b>Permissions Required</b>: Manage Patients'
      externalDocs:
        x-doc-key: get_patients_v2_active
      operationId: GetPatientActive
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to return the list of active patients.
        format: ''
      - type: integer
        name: pageNumber
        in: query
        description: The number of the data page from which to return data.
        format: ''
      - type: integer
        name: pageSize
        in: query
        description: The number of records to return per page. Pagination is currently disabled by default. You can enable pagination on this query by specifying a value that does not exceed 20.
        format: ''
      responses:
        200:
          description: Ok
  /patients/v2/:
    post:
      tags:
      - Patients
      summary: ''
      description: '<b>Permissions Required</b>: Manage Patients'
      externalDocs:
        x-doc-key: post_patients_v2_
      operationId: PostPatients
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to record the list of patients.
        format: ''
      responses:
        200:
          description: Ok
    put:
      tags:
      - Patients
      summary: ''
      description: '<b>Permissions Required</b>: Manage Patients'
      externalDocs:
        x-doc-key: put_patients_v2_
      operationId: PutPatients
      parameters:
      - type: string
        name: licenseNumber
        in: query
        required: true
        description: The license number of the facility for which to update the list of patients.
        format: ''
      responses:
        200:
          description: Ok
...